feat(meal-plan): add effective portions and group member management
- Introduced `effective_portions` to support float values for `norm_portions` in meal plans.
- Implemented CRUD operations for `MealPlanGroupMember`, including validation and automatic recalculation of `norm_portions`.
- Enhanced meal plan detail responses to include group members and their attributes.
- Added bulk creation endpoint for group members and synchronization with event participants.
- Updated `norm_portions` to be derived from group members automatically, transitioning to a float field.
- Implemented activity factor management in meal plans, affecting group member calculations.
- Added frontend components for managing group members, including forms and dialogs for creation and synchronization.
- Introduced new ingredient creation workflows with query parameter support for pre-filling and redirection.
- Enhanced recipe AI creation and suggestion features with new endpoints and validation schemas.

+++ backend/planner/services/cooking_schedule_service.py
@@ -150,7 +150,7 @@ class CookingScheduleResult:
     total_cost_eur: float = 0.0
     total_cost_with_reserve: float = 0.0
     total_energy_kcal: float = 0.0
-    norm_portions: int = 0
+    norm_portions: float = 0.0
 
 
 def parse_recipe_steps(markdown: str) -> list[CookingScheduleStep]:
